Add requested formatter module
