+++ lib/bgs/dependents.rb
@@ -2,14 +2,7 @@
 
 module BGS
   class Dependents < Service
-    CHILD_STATUS = {
-      'child_under18' => 'Other',
-      'step_child' => 'Stepchild',
-      'biological' => 'Biological',
-      'adopted' => 'Adopted Child',
-      'disabled' => 'Other',
-      'child_over18_in_school' => 'Other'
-    }.freeze
+    include Helpers::Formatters
 
     def initialize(proc_id:, payload:, user:)
       @proc_id = proc_id
@@ -33,6 +26,7 @@ module BGS
 
       @dependents
     end
+
     # rubocop:enable Metrics/CyclomaticComplexity
     # rubocop:enable Metrics/PerceivedComplexity
 
@@ -83,7 +77,7 @@ module BGS
     end
 
     def report_divorce
-      divorce_info = format_divorce_info
+      divorce_info = format_divorce_info(@dependents_application['report_divorce'])
       participant = create_participant(@proc_id)
       create_person(@proc_id, participant[:vnp_ptcpnt_id], divorce_info)
 
@@ -119,7 +113,7 @@ module BGS
     end
 
     def report_child_event(event_type)
-      formatted_child_info = format_child_event(event_type)
+      formatted_child_info = child_event_info(event_type)
       participant = create_participant(@proc_id)
 
       create_person(@proc_id, participant[:vnp_ptcpnt_id], formatted_child_info)
@@ -135,14 +129,11 @@ module BGS
       )
     end
 
-    def format_child_event(event_type)
-      return format_child_marriage_info(@dependents_application['child_marriage']) if event_type == 'child_marriage'
-
-      format_child_not_attending(@dependents_application['child_stopped_attending_school'])
-    end
-
     def report_674
-      formatted_674_info = format_674_info
+      formatted_674_info = format_674_info(
+        @dependents_application['student_name_and_ssn'],
+        @dependents_application.dig('student_address_marriage_tuition', 'was_married')
+      )
       student_address = @dependents_application['student_address_marriage_tuition']['address']
       participant = create_participant(@proc_id)
       create_person(@proc_id, participant[:vnp_ptcpnt_id], formatted_674_info)
@@ -156,74 +147,10 @@ module BGS
       )
     end
 
-    def format_674_info
-      name_and_ssn = @dependents_application['student_name_and_ssn']
-      was_married = @dependents_application.dig('student_address_marriage_tuition', 'was_married')
-      {
-        'ssn': name_and_ssn['ssn'],
-        'birth_date': name_and_ssn['birth_date'],
-        'ever_married_ind': was_married == true ? 'Y' : 'N'
-      }.merge(name_and_ssn['full_name']).with_indifferent_access
-    end
+    def child_event_info(event_type)
+      return format_child_marriage_info(@dependents_application['child_marriage']) if event_type == 'child_marriage'
 
-    def format_child_marriage_info(child_marriage)
-      {
-        'event_date': child_marriage['date_married']
-      }.merge(child_marriage['full_name']).with_indifferent_access
-    end
-
-    def format_child_not_attending(child)
-      {
-        event_date: child['date_child_left_school']
-      }.merge(child['full_name']).with_indifferent_access
-    end
-
-    def format_stepchild_info(stepchild_info)
-      {
-        'living_expenses_paid': stepchild_info['living_expenses_paid'],
-        'lives_with_relatd_person_ind': 'N'
-      }.merge(stepchild_info['full_name']).with_indifferent_access
-    end
-
-    def format_child_info(child_info)
-      {
-        'ssn': child_info['ssn'],
-        'family_relationship_type': CHILD_STATUS[child_info['child_status'].key(true)],
-        'place_of_birth_state': child_info.dig('place_of_birth', 'state'),
-        'place_of_birth_city': child_info.dig('place_of_birth', 'city'),
-        'reason_marriage_ended': child_info.dig('previous_marriage_details', 'reason_marriage_ended'),
-        'ever_married_ind': child_info['previously_married'] == 'Yes' ? 'Y' : 'N'
-      }.merge(child_info['full_name']).with_indifferent_access
-    end
-
-    def format_death_info(death_info)
-      {
-        'death_date': death_info['date'],
-        'vet_ind': 'N'
-      }.merge(death_info['full_name'])
-    end
-
-    def format_divorce_info
-      report_divorce = @dependents_application['report_divorce']
-      {
-        divorce_state: report_divorce.dig('location', 'state'),
-        divorce_city: report_divorce.dig('location', 'city'),
-        marriage_termination_type_code: report_divorce['reason_marriage_ended'],
-        event_dt: report_divorce['date'],
-        vet_ind: 'N',
-        type: 'divorce'
-      }.merge(report_divorce['full_name']).with_indifferent_access
-    end
-
-    def relationship_type(info)
-      if info['dependent_type']
-        return { participant: 'Guardian', family: 'Other' } if info['dependent_type'] == 'DEPENDENT_PARENT'
-
-        {
-          participant: info['dependent_type'].capitalize.gsub('_', ' '),
-          family: info['dependent_type'].capitalize.gsub('_', ' ')
-        }
-      end
+      format_child_not_attending(@dependents_application['child_stopped_attending_school'])
     end
   end
 end
